\name{dim-methods}
\alias{dim.QRectF}
\title{Get the dimensions of rectangles and rectangular objects...}
\usage{\method{dim}{QRectF}(x)
\method{dim}{QGraphicsScene}(x)
\method{dim}{QGraphicsItem}(x)
\method{dim}{QGraphicsView}(x)
}
\description{Get the dimensions of rectangles and rectangular objects}
\arguments{\item{x}{A rectangular object, like \code{QRect} or \code{QGraphicsItem}}
}
\alias{dim.QRect}
\alias{dim.QGraphicsScene}
\alias{dim.QGraphicsItem}
\alias{dim.QGraphicsView}

